Fix UI, clean up camera/selection, comments, bugs fixed (#7)
* Fix UI, clean up camera and selection

Made camera code less bulky by using an array, moved selection code to proper place preventing the need for a redraw variable

* Added functionality to selected items label

* Added comments/fixed capitalization

Fixed capitalization in a few spots and added some more clarification for the variables at the beginning.

* Fixed selection code

Noticed that if LeftShift is being held and the user clicks on a part to deselect it, it deselects but immediately reselects. Moved line to select mouseHit to the correct spot.