Simplify calculateAngleBetween3Points
Seen as a,b,c were entered in as raw values to begin with, I don't think that declaring operations (a-b) & (c-b) is/are sufficient as variables. Should probably check to see if a,b,c exists before it's being passed in to prevent possible edge cases from happening. 

Probably a good idea to implement error (exits, but provides a message) and warn (pass, but provides a message) methods or something similar.